refactor(tests): use common 'mock_document_service' in unittests
