+++ src/tests/unit/mock_utils.py
@@ -1,7 +1,11 @@
+from collections import defaultdict
+
+from domain_classes.storage_recipe import StorageAttribute, StorageRecipe
+from enums import SIMOS, StorageDataTypes
+from services.document_service import DocumentService
+
+
 # https://stackoverflow.com/questions/6027558/flatten-nested-dictionaries-compressing-keys
-from enums import SIMOS
-
-
 def flatten_dict(dd, separator="_", prefix=""):
     return (
         {
@@ -281,3 +285,94 @@ class BlueprintProvider:
 
 
 blueprint_provider = BlueprintProvider()
+
+all_storage_recipes = defaultdict(list)
+all_storage_recipes.update(
+    {
+        "all_contained_cases_blueprint": [
+            {
+                "type": "dmss://system/SIMOS/StorageRecipe",
+                "name": "DefaultStorageRecipe",
+                "description": "",
+                "attributes": [
+                    {"name": "reference", "type": "dmss://system/SIMOS/Entity", "contained": False},
+                    {"name": "references", "type": "dmss://system/SIMOS/Entity", "contained": False},
+                ],
+            }
+        ],
+        "uncontained_blueprint": [
+            {
+                "type": "dmss://system/SIMOS/StorageRecipe",
+                "name": "DefaultStorageRecipe",
+                "description": "",
+                "attributes": [
+                    {
+                        "name": "uncontained_in_every_way",
+                        "type": "does_this_matter?",
+                        "contained": False,
+                        "storageAffinity": "blob",
+                    }
+                ],
+            }
+        ],
+        "ExtendedBlueprint": [
+            {
+                "name": "default",
+                "type": "dmss://system/SIMOS/StorageRecipe",
+                "description": "",
+                "storageAffinity": "blob",
+                "attributes": [],
+            }
+        ],
+        # "uncontained_blueprint": uncontained_blueprint,
+        "blob": [
+            {
+                "name": "default",
+                "type": "dmss://system/SIMOS/StorageRecipe",
+                "description": "",
+                "storageAffinity": "blob",
+                "attributes": [],
+            }
+        ],
+        "blobContainer": [
+            {
+                "name": "default",
+                "type": "dmss://system/SIMOS/StorageRecipe",
+                "description": "",
+                "attributes": [{"name": "blob", "type": "does_this_matter?", "contained": False}],
+            }
+        ],
+    }
+)
+
+
+def mock_storage_recipe_provider(type: str, context: str) -> list[StorageRecipe]:
+    return [
+        StorageRecipe(
+            name=sr.get("name"),
+            storage_affinity=StorageDataTypes(sr.get("storageAffinity", "default")),
+            attributes={
+                a["name"]: StorageAttribute(
+                    name=a["name"], contained=a["contained"], storage_affinity=a.get("storageAffinity", "default")
+                )
+                for a in sr.get("attributes", [])
+            },
+        )
+        for sr in all_storage_recipes[type]
+    ]
+
+
+def get_mock_document_service(
+    repository_provider=None,
+    blueprint_provider=blueprint_provider,
+    recipe_provider=mock_storage_recipe_provider,
+    user=None,
+    context=None,
+):
+    return DocumentService(
+        blueprint_provider=blueprint_provider,
+        recipe_provider=recipe_provider,
+        repository_provider=repository_provider,
+        user=user,
+        context=context,
+    )
