+++ src/tests/unit/test_save.py
@@ -5,9 +5,11 @@ from unittest import mock
 from authentication.models import User
 from domain_classes.blueprint_attribute import BlueprintAttribute
 from domain_classes.tree_node import Node
-from services.document_service import DocumentService
-from tests.unit.mock_blueprint_provider import blueprint_provider, flatten_dict
-from tests.unit.mock_storage_recipe_provider import mock_storage_recipe_provider
+from tests.unit.mock_utils import (
+    flatten_dict,
+    get_mock_document_service,
+    mock_storage_recipe_provider,
+)
 
 
 class DocumentServiceTestCase(unittest.TestCase):
@@ -41,11 +43,7 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=blueprint_provider,
-            repository_provider=lambda x, y: repository,
-        )
+        document_service = get_mock_document_service(lambda x, y: repository)
 
         node: Node = document_service.get_node_by_uid("testing", "1")
         contained_node: Node = node.get_by_path("references.1".split("."))
@@ -85,11 +83,7 @@ class DocumentServiceTestCase(unittest.TestCase):
         repository.get = mock_get
         repository.update = mock_update
 
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=blueprint_provider,
-            repository_provider=lambda x, y: repository,
-        )
+        document_service = get_mock_document_service(lambda x, y: repository)
 
         node: Node = document_service.get_node_by_uid("testing", "1")
         contained_node: Node = node.search("1.references")
@@ -157,11 +151,7 @@ class DocumentServiceTestCase(unittest.TestCase):
             if data_source_id == "testing":
                 return repository
 
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=blueprint_provider,
-            repository_provider=repository_provider,
-        )
+        document_service = get_mock_document_service(repository_provider)
 
         node: Node = document_service.get_node_by_uid("testing", "1")
         contained_node: Node = node.search("1.references")
@@ -206,11 +196,7 @@ class DocumentServiceTestCase(unittest.TestCase):
         repository.get = lambda id: doc_storage[id]
         repository.update = mock_update
 
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=blueprint_provider,
-            repository_provider=lambda x, y: repository,
-        )
+        document_service = get_mock_document_service(lambda x, y: repository)
 
         node: Node = Node.from_dict(
             doc_storage["1"], "1", document_service.get_blueprint, recipe_provider=mock_storage_recipe_provider
@@ -264,11 +250,7 @@ class DocumentServiceTestCase(unittest.TestCase):
         repository.get = lambda id: deepcopy(doc_storage[id])
         repository.update = mock_update
 
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=blueprint_provider,
-            repository_provider=lambda x, y: repository,
-        )
+        document_service = get_mock_document_service(lambda x, y: repository)
 
         document_service.update_document(
             "test",
@@ -328,11 +310,7 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=blueprint_provider,
-            repository_provider=lambda x, y: repository,
-        )
+        document_service = get_mock_document_service(lambda x, y: repository)
         new_data = {
             "name": "A-contained_attribute",
             "type": "all_contained_cases_blueprint",
