+++ src/tests/unit/test_tree_node_update.py
@@ -7,10 +7,8 @@ from common.utils.data_structure.compare import pretty_eq
 from domain_classes.blueprint import Blueprint
 from domain_classes.tree_node import Node
 from enums import SIMOS
-from services.document_service import DocumentService
 from storage.repositories.file import LocalFileRepository
-from tests.unit.mock_blueprint_provider import blueprint_provider
-from tests.unit.mock_storage_recipe_provider import mock_storage_recipe_provider
+from tests.unit.mock_utils import get_mock_document_service
 
 
 class MultiTypeBlueprintProvider:
@@ -155,11 +153,7 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            blueprint_provider=blueprint_provider,
-            repository_provider=repository_provider,
-            recipe_provider=mock_storage_recipe_provider,
-        )
+        document_service = get_mock_document_service(repository_provider)
 
         node: Node = document_service.get_node_by_uid("testing", "1")
         node.update(
@@ -209,11 +203,7 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            blueprint_provider=blueprint_provider,
-            repository_provider=repository_provider,
-            recipe_provider=mock_storage_recipe_provider,
-        )
+        document_service = get_mock_document_service(repository_provider)
         document_service.add_document(
             "testing/1.im_optional",
             data={"type": "basic_blueprint", "name": "new_entity", "description": "This is my new entity"},
@@ -243,10 +233,8 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            blueprint_provider=MultiTypeBlueprintProvider(),
-            repository_provider=lambda x, y: repository,
-            recipe_provider=mock_storage_recipe_provider,
+        document_service = get_mock_document_service(
+            blueprint_provider=MultiTypeBlueprintProvider(), repository_provider=lambda x, y: repository
         )
         with self.assertRaises(BadRequestException):
             document_service.update_document(
@@ -304,11 +292,7 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            blueprint_provider=blueprint_provider,
-            repository_provider=repository_provider,
-            recipe_provider=mock_storage_recipe_provider,
-        )
+        document_service = get_mock_document_service(repository_provider)
         document_service.add_document(
             "testing/1.nested_with_optional.im_optional",
             {"name": "new_entity", "description": "This is my new entity", "type": "basic_blueprint"},
@@ -342,11 +326,7 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            blueprint_provider=blueprint_provider,
-            repository_provider=lambda x, y: repository,
-            recipe_provider=mock_storage_recipe_provider,
-        )
+        document_service = get_mock_document_service(lambda x, y: repository)
 
         with self.assertRaises(BadRequestException):
             document_service.add_document(
@@ -367,10 +347,8 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=MultiTypeBlueprintProvider(),
-            repository_provider=lambda x, y: repository,
+        document_service = get_mock_document_service(
+            lambda id, user: repository, blueprint_provider=MultiTypeBlueprintProvider()
         )
         document_service.update_document(
             data_source_id="testing",
@@ -397,10 +375,8 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=MultiTypeBlueprintProvider(),
-            repository_provider=lambda x, y: repository,
+        document_service = get_mock_document_service(
+            lambda id, user: repository, blueprint_provider=MultiTypeBlueprintProvider()
         )
         document_service.update_document(
             data_source_id="testing",
@@ -436,10 +412,8 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=MultiTypeBlueprintProvider(),
-            repository_provider=lambda x, y: repository,
+        document_service = get_mock_document_service(
+            lambda id, user: repository, blueprint_provider=MultiTypeBlueprintProvider()
         )
         document_service.update_document(
             data_source_id="testing",
@@ -481,10 +455,8 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=MultiTypeBlueprintProvider(),
-            repository_provider=lambda x, y: repository,
+        document_service = get_mock_document_service(
+            lambda id, user: repository, blueprint_provider=MultiTypeBlueprintProvider()
         )
 
         with self.assertRaises(BadRequestException) as error:
@@ -518,10 +490,8 @@ class DocumentServiceTestCase(unittest.TestCase):
 
         repository.get = mock_get
         repository.update = mock_update
-        document_service = DocumentService(
-            recipe_provider=mock_storage_recipe_provider,
-            blueprint_provider=MultiTypeBlueprintProvider(),
-            repository_provider=lambda x, y: repository,
+        document_service = get_mock_document_service(
+            lambda id, user: repository, blueprint_provider=MultiTypeBlueprintProvider()
         )
 
         document_service.update_document(
